+++ src/Akeneo/AssetManager/back/Infrastructure/Persistence/Sql/Subscribers/RefreshAssetsSubscriber.php
@@ -4,9 +4,8 @@ declare(strict_types=1);
 namespace Akeneo\AssetManager\Infrastructure\Persistence\Sql\Subscribers;
 
 use Akeneo\AssetManager\Domain\Event\AssetDeletedEvent;
-use Akeneo\AssetManager\Domain\Event\AssetFamilyAssetsDeletedEvent;
+use Akeneo\AssetManager\Domain\Event\AssetsDeletedEvent;
 use Akeneo\AssetManager\Domain\Event\AttributeOptionsDeletedEvent;
-use Akeneo\AssetManager\Infrastructure\Persistence\Sql\CLI\RefreshAssetsCommand;
 use Akeneo\Tool\Component\Console\CommandLauncher;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -32,7 +31,7 @@ class RefreshAssetsSubscriber implements EventSubscriberInterface
 //        To activate when we improve the refresh performances
 //        return [
 //            AssetDeletedEvent::class => 'onEvent',
-//            AssetFamilyAssetsDeletedEvent::class => 'onEvent',
+//            AssetsDeletedEvent::class => 'onEvent',
 //            AttributeOptionsDeletedEvent::class => 'onEvent',
 //        ];
         return [];